fix(auth): second round of cubic findings on the callback's error path
- clear the one-shot login-state cookie on the provider-denial path too,
  not just the other five refusal paths — it was otherwise left usable
  for its full 10-minute TTL.
- bound the Google `error` value logged to system_log: it comes straight
  off an unauthenticated callback's query string.
- rewrite the split-origin comment in cmd/api/googlesignin.go: it drew a
  distinction from connectors.go's own APIBaseURL usage that isn't
  actually there (both build an outbound redirect_uri the same way).
- fix the migration comment misattributing resolveFederatedUser's
  (provider, subject) lookup to the (user_id, provider) index.
- rewrite the provider-denial test so it can't stay green by accident:
  a verifier/exchanger/state signer that would all otherwise complete
  the round trip, so `error` is the only thing standing between the
  request and a session.
- split OIDCRoutes/callbackURI/WithOIDCProviders/WithOIDCCapabilitiesFn
  into oidcroutes.go — ssologin.go had grown past the 500-line ceiling.
